feat: rename resouces api routes
