+++ src/app/api/resources/route.ts
@@ -1,18 +1,13 @@
 import { Resource } from "@/app/schema"
 import { DataSource } from "@/services/datasource"
-import { NextResponse, NextRequest } from "next/server"
+import { NextResponse } from "next/server"
 
-export async function GET(req: NextRequest) {
-    const resourceId = req.nextUrl.searchParams.get('resource_id')
-  
-    if(!resourceId) {
-      return NextResponse.json({ error: 'El id del recurso es requerido' }, { status: 400 })
-    }
-    try { 
-      const response: Resource[] = await new DataSource().getResource(resourceId)
-      return NextResponse.json({ message: `Recurso ${response[0].name} obtenido`, data: response })
-    } catch (error) {
-      console.error('Error en /api/register:', (error as Error).message)
-      return NextResponse.json({ error: `No se pudo obtener el recurso. Intentá más tarde. Error: ${(error as Error).message}` }, { status: 500 })
-    }
+export async function GET() {
+  try {
+    const response: Resource[] = await new DataSource().getAllResources();
+    return NextResponse.json({ message: 'Todos los recursos obtenidos', data: response, status: 200 });
+  } catch (error) {
+    console.error('Error en /api/resources:', (error as Error).message);
+    return NextResponse.json({ error: `No se pudieron obtener los recursos. Intentá más tarde. Error: ${(error as Error).message}` }, { status: 500 });
   }
+}
